More attempts at debugging. Issue in description.
In brief, the TWDR we are checking for the address gets overwritten with new data when the Arduino is being written to. Using this method worked just fine when having data requested since no new data would be coming in from the master, but when data is being sent our handling code isn't called until the full message is received. This handling occurs in twi.c